Refactor importers to package
